requirements.txt
# Pip to install

django==3.2.3
django-inline-svg==0.1.1
djangorestframework==3.12.4
django-svg-templatetag==1.0.1
social-auth-app-django==4.0.0
geocoder==1.38.1
django-credit-cards==0.4.1
django-crispy_forms==1.12.0
mysqlclient==2.0.3
